Add navigation module
