+++ index.html
@@ -11,9 +11,9 @@
 <header class="header">
     <a href="#">Awesome Books</a>
     <nav class="nav">
-        <a onclick="navSection('booklist')" class="list-nav">List</a>
-        <a onclick="navSection('addBook')" class="add-nav">Add new</a>
-        <a onclick="navSection('contact')" class="contact-nav">Contact</a>
+        <a  class="list-nav">List</a>
+        <a  class="add-nav">Add new</a>
+        <a  class="contact-nav">Contact</a>
     </nav>
 </header>
 <section class="books" id="books">
